% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_effects.R
\name{capture_side_effects}
\alias{capture_side_effects}
\title{Capture side effects}
\usage{
capture_side_effects(fn, ...)
}
\arguments{
\item{fn}{Function to evaluate.}

\item{...}{Arguments for \code{fn}.}
}
\value{
Named list with the side effects.
}
\description{
Captures errors, warnings, and messages from running a given function.

When the function throws an error, no other side effects are captured.

Simple wrapper for testthat's
\code{\link[testthat:capture_error]{capture_error()}},
\code{\link[testthat:capture_warnings]{capture_warnings()}} and
\code{\link[testthat:capture_messages]{capture_messages()}}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
