% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_test_that_addin.R
\name{initializeTestthatAddin}
\alias{initializeTestthatAddin}
\title{Initializes \code{test_that()} call}
\usage{
initializeTestthatAddin(insert = TRUE, indentation = NULL)
}
\arguments{
\item{insert}{Whether to insert the code via
\code{\link[rstudioapi:insertText]{rstudioapi::insertText()}}
or return it. (Logical)

\strong{N.B.} Mainly intended for testing the addin programmatically.}

\item{indentation}{Indentation of the code. (Numeric)

\strong{N.B.} Mainly intended for testing the addin programmatically.}
}
\value{
Inserts code for calling
\code{\link[testthat:test_that]{testthat::test_that()}}.

Returns \code{NULL} invisibly.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Inserts code for calling \code{\link[testthat:test_that]{testthat::test_that()}}.

See \code{Details} for how to set a key command.
}
\details{
\subsection{How to set up a key command in RStudio}{

After installing the package.
Go to:

\code{Tools >> Addins >> Browse Addins >> Keyboard Shortcuts}.

Find \code{"Initialize test_that()"} and press its field under \code{Shortcut}.

Press desired key command, e.g. \code{Alt+T}.

Press \code{Apply}.

Press \code{Execute}.
}
}
\seealso{
Other addins: 
\code{\link{assertCollectionAddin}()},
\code{\link{dputSelectedAddin}()},
\code{\link{initializeGXSFunctionAddin}()},
\code{\link{insertExpectationsAddin}()},
\code{\link{wrapStringAddin}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{addins}
