% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_string_addin.R
\name{wrapStringAddin}
\alias{wrapStringAddin}
\title{Wraps the selection with paste0}
\usage{
wrapStringAddin(selection = NULL, insert = TRUE, indentation = 0)
}
\arguments{
\item{selection}{String of code. (Character)

N.B. Mainly intended for testing the addin programmatically.}

\item{insert}{Whether to insert the wrapped text via
\code{\link[rstudioapi:insertText]{rstudioapi::insertText()}}
or return it. (Logical)

N.B. Mainly intended for testing the addin programmatically.}

\item{indentation}{Indentation of the selection. (Numeric)

N.B. Mainly intended for testing the addin programmatically.}
}
\value{
Inserts \code{paste0("first n chars", "next n chars")} etc.

Returns \code{NULL} invisibly.
}
\description{
Splits the selection every n characters
and inserts it in a \code{\link[base:paste0]{paste0()}} call.

See \code{Details} for how to set a key command.
}
\details{
\subsection{How to set up a key command in RStudio}{

After installing the package.
Go to:

\code{Tools >> Addins >> Browse Addins >> Keyboard Shortcuts}.

Find \code{"Insert Expectations"} and press its
field under \code{Shortcut}.

Press desired key command, e.g. \code{Alt+P}.

Press \code{Apply}.

Press \code{Execute}.
}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
