% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{xpectr-package}
\alias{xpectr}
\alias{xpectr-package}
\title{xpectr: A package for generating tests for \code{testthat} unit testing}
\description{
A set of utilities and RStudio addins for generating tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ludvigolsen/xpectr}
  \item Report bugs at \url{https://github.com/ludvigolsen/xpectr/issues}
}

}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
