% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{xpectr}
\alias{xpectr}
\title{xpectr: A package for generating tests for \code{testthat} unit testing}
\description{
A set of utilities and RStudio addins for generating tests.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
