% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_expectations_from_selection.R
\name{gxs_selection}
\alias{gxs_selection}
\title{Generate testhat expectations from selection}
\usage{
gxs_selection(
  selection,
  indentation = 0,
  strip = TRUE,
  tolerance = "1e-4",
  envir = NULL,
  out = "insert"
)
}
\arguments{
\item{selection}{String of code. (Character)

E.g. \code{"stop('This gives an expect_error test')"}.}

\item{indentation}{Indentation of the selection. (Numeric)}

\item{strip}{Whether to insert
\code{\link[xpectr:strip]{strip_msg()}} and
\code{\link[xpectr:strip]{strip()}}
in tests of side effects. (Logical)

Sometimes testthat tests have differences in punctuation and newlines on different
systems. By stripping both the error message and the expected message of non-alphanumeric symbols,
we can avoid such failed tests.}

\item{tolerance}{The tolerance for numeric tests as a string. (Character)}

\item{out}{Either "insert" or "return".

\subsection{"insert" (Default)}{
Inserts the expectations via
\code{\link[rstudioapi:insertText]{rstudioapi::insertText()}}.
}
\subsection{"return"}{
Returns the expectations in a list.

These can be prepared for insertion with
\code{\link[xpectr:prepare_insertion]{prepare_insertion()}}.
}}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Based on the selection (string of code), a set of testthat \code{expect_*}
statements are generated.

Example: If the selected code is the name of a data frame object,
it will create an \code{\link[testthat:expect_equal]{expect_equal}}
test for each column, along with a test of the column names.

Currently supports side effects (error, warnings, messages),
data frames, and vectors.

List columns in data frames (like nested tibbles) are currently skipped.

Addin: \code{\link[xpectr:insertExpectationsAddin]{insertExpectationsAddin()}}
}
\examples{
# Attach packages
library(xpectr)

\donttest{
df <- data.frame('a' = c(1, 2, 3), 'b' = c('t', 'y', 'u'),
                 stringsAsFactors = FALSE)

gxs_selection("stop('This gives an expect_error test!')")
gxs_selection("warning('This gives an expect_warning test!')")
gxs_selection("message('This gives an expect_message test!')")
gxs_selection("stop('This: tests the -> punctuation!')", strip = FALSE)
gxs_selection("sum(1, 2, 3, 4)")
gxs_selection("df")

tests <- gxs_selection("df", out = "return")
for_insertion <- prepare_insertion(tests)
rstudioapi::insertText(for_insertion)
}
}
\seealso{
Other expectation generators: 
\code{\link{insertExpectationsAddin}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{expectation generators}
